# doc/openapi.yml
openapi: 3.1.0
info:
  title: Card Collection Price Tracker API
  description: 'REST API for tracking Magic: The Gathering card prices'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 0.1.0
paths:
  /cards/card-info:
    post:
      tags:
      - cards
      operationId: get_card_info
      responses:
        '200':
          description: Card info from EDHRec
        '401':
          description: Missing or invalid token
      security:
      - bearer_auth: []
  /cards/import:
    post:
      tags:
      - cards
      operationId: import_cards
      requestBody:
        description: ManaBox CSV content (max 10 MB)
        content:
          text/plain:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid body (non UTF-8, ...)
        '401':
          description: Missing or invalid token
      security:
      - bearer_auth: []
  /cards/{scryfall_id}/price-history:
    get:
      tags:
      - cards
      operationId: get_card_price_history
      parameters:
      - name: scryfall_id
        in: path
        description: Card's Scryfall identifier
        required: true
        schema:
          type: string
          format: uuid
      - name: start_date
        in: query
        description: 'Start date (ISO 8601: YYYY-MM-DD, inclusive). Defaults to end_date minus 30 days'
        required: false
        schema:
          type: string
      - name: end_date
        in: query
        description: 'End date (ISO 8601: YYYY-MM-DD, inclusive). Defaults to today'
        required: false
        schema:
          type: string
      responses:
        '200':
          description: Card price history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PriceHistoryEntryResponse'
        '400':
          description: Invalid date range (start_date > end_date)
        '401':
          description: Missing or invalid token
        '404':
          description: No card found for this scryfall_id
      security:
      - bearer_auth: []
  /collection:
    get:
      tags:
      - collection
      operationId: get_collection
      parameters:
      - name: page
        in: query
        description: Page number (starts at 0)
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: page_size
        in: query
        description: Items per page (max 100)
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: sort_by
        in: query
        description: Sort field
        required: false
        schema:
          $ref: '#/components/schemas/SortByParam'
      - name: sort_dir
        in: query
        description: Sort direction
        required: false
        schema:
          $ref: '#/components/schemas/SortDirParam'
      - name: q
        in: query
        description: Fuzzy search on card name or set
        required: false
        schema:
          type: string
      - name: rarity
        in: query
        description: Comma-separated rarity codes (C, U, R, M)
        required: false
        schema:
          type: string
      - name: sets
        in: query
        description: Comma-separated set codes
        required: false
        schema:
          type: string
      - name: price_min
        in: query
        description: Minimum trend price in cents
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: price_max
        in: query
        description: Maximum trend price in cents
        required: false
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: owned
        in: query
        description: 'Restrict to cards owned by the authenticated user (default: false — full catalog)'
        required: false
        schema:
          type: boolean
      responses:
        '200':
          description: Paginated card collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCollectionResponse'
        '401':
          description: Missing or invalid token
      security:
      - bearer_auth: []
  /maintenance/stats:
    get:
      tags:
      - maintenance
      operationId: get_stats
      responses:
        '200':
          description: Global database statistics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatsResponse'
  /maintenance/trigger-price-update:
    post:
      tags:
      - maintenance
      operationId: trigger_price_update
      responses:
        '204':
          description: Price update triggered successfully
  /maintenance/update-cardmarket-ids:
    post:
      tags:
      - maintenance
      operationId: update_cardmarket_ids
      responses:
        '202':
          description: CardMarket IDs enqueued for update
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnqueueResponse'
  /trades:
    post:
      tags:
      - trades
      operationId: create_trade
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTradeRequest'
        required: true
      responses:
        '201':
          description: Trade created, or card added to (merged into) the active trade with this user
        '400':
          description: Invalid payload or respondent is the initiator
        '401':
          description: Missing or invalid token
        '404':
          description: Card not found, not owned by respondent, or respondent unknown
        '409':
          description: The active trade with this user is already fully accepted and can no longer be modified
      security:
      - bearer_auth: []
  /user/register:
    post:
      tags:
      - auth
      operationId: register
      responses:
        '204':
          description: User registered/updated successfully
        '400':
          description: Missing username claim in token
        '401':
          description: Missing or invalid authentication token
      security:
      - bearer_auth: []
components:
  schemas:
    CollectionCardResponse:
      type: object
      required:
      - set_code
      - collector_number
      - language_code
      - foil
      - name
      - rarity_code
      - scryfall_id
      properties:
        collection_entry:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/CollectionEntryResponse'
            description: Present only when the card is owned by the authenticated user.
        collector_number:
          type: string
        foil:
          type: boolean
        language_code:
          type: string
        name:
          type: string
        owner_username:
          type:
          - string
          - 'null'
          description: Username of the owner when the card belongs to another user (catalog listing).
        price_guide:
          oneOf:
          - type: 'null'
          - $ref: '#/components/schemas/PriceGuideResponse'
        rarity_code:
          type: string
        scryfall_id:
          type: string
        set_code:
          type: string
        the_gatherer_id:
          type:
          - string
          - 'null'
    CollectionEntryResponse:
      type: object
      required:
      - quantity
      - purchase_price
      - added_at
      properties:
        added_at:
          type: string
          description: RFC 3339 timestamp
        purchase_price:
          type: integer
          format: int32
          minimum: 0
        quantity:
          type: integer
          format: int32
          minimum: 0
    CreateTradeRequest:
      type: object
      required:
      - set_code
      - collector_number
      - language_code
      - foil
      - respondent_user_id
      - quantity
      properties:
        collector_number:
          type: string
        foil:
          type: boolean
        language_code:
          type: string
        quantity:
          type: integer
          format: int32
          minimum: 0
        respondent_user_id:
          type: string
        set_code:
          type: string
    EnqueueResponse:
      type: object
      required:
      - enqueued
      properties:
        enqueued:
          type: integer
          minimum: 0
    MessageResponse:
      type: object
      required:
      - message
      properties:
        message:
          type: string
    PaginatedCollectionResponse:
      type: object
      required:
      - items
      - total
      - page
      - page_size
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CollectionCardResponse'
        page:
          type: integer
          format: int32
          minimum: 0
        page_size:
          type: integer
          format: int32
          minimum: 0
        total:
          type: integer
          format: int64
          minimum: 0
    PriceGuideResponse:
      type: object
      properties:
        avg:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        low:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
        trend:
          type:
          - integer
          - 'null'
          format: int32
          minimum: 0
    PriceHistoryEntryResponse:
      type: object
      required:
      - date
      - low
      - trend
      - avg
      properties:
        avg:
          type: integer
          format: int64
        date:
          type: string
          description: ISO 8601 date string (YYYY-MM-DD)
        low:
          type: integer
          format: int64
        trend:
          type: integer
          format: int64
    SortByParam:
      type: string
      enum:
      - avg
      - trend
      - set_code
      - language_code
    SortDirParam:
      type: string
      enum:
      - asc
      - desc
    StatsResponse:
      type: object
      required:
      - card_number
      - card_price_number
      - db_size_mb
      properties:
        card_number:
          type: integer
          format: int32
          minimum: 0
        card_price_number:
          type: integer
          format: int32
          minimum: 0
        db_size_mb:
          type: integer
          format: int32
          minimum: 0
  securitySchemes:
    bearer_auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
tags:
- name: cards
  description: Card catalog, import and price history (authentication required)
- name: collection
  description: Player's collection (authentication required)
- name: maintenance
  description: Maintenance operations (public)
- name: auth
  description: Authentication and user registration (authentication required)
- name: trades
  description: Trade requests between two collectors (authentication required)